% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{model_url}
\alias{model_url}
\title{Model endpoint URLs}
\usage{
model_url(model = c("qwen3-8b", "zeta"))
}
\arguments{
\item{model}{Character. One of \code{"zeta"} or \code{"qwen3-8b"}.}
}
\value{
A single character string: the endpoint URL.
}
\description{
Retrieve the endpoint URL for a given model. URLs are read from environment
variables so that they do not need to be stored in the package source.

The expected environment variables are:
\itemize{
\item \code{NESEVALS_ZETA_URL} -- endpoint URL for the Zeta model
\item \code{NESEVALS_QWEN3_8B_URL} -- endpoint URL for the Qwen3-8B model
}

A convenient place to set these is your \code{.Renviron} file (see
\code{\link[usethis:edit]{usethis::edit_r_environ()}}).

The endpoints will go to sleep after 15 minutes. As you begin working on
features that will use these models, you may want to ping the endpoints
to wake the instance up (it will take a few minutes).
}
\examples{
\dontrun{
model_url("zeta")
model_url("qwen3-8b")
}

}
